Move existing auth and files into access
