+++ api-v2/app/access/auth.py
@@ -8,9 +8,9 @@ from sqlalchemy.orm import Session
 from botocore.exceptions import ClientError
 
 
-from app.schemas import UserCreate, UserSignIn, UserSignInResponse, ForgotPasswordRequest, ForgotPasswordResponse, ConfirmForgotPasswordResponse, ConfirmForgotPasswordRequest
+from app.access.schemas import UserCreate, UserSignInRequest, UserSignInResponse, ForgotPasswordRequest, ForgotPasswordResponse, ConfirmForgotPasswordResponse, ConfirmForgotPasswordRequest, RefreshTokenResponse
 
-from app.crud import create_user, delete_user, get_user
+from app.access.crud import create_user, delete_user, get_user
 from app.api.deps import (
     get_db,
     get_cognito_client,
