+++ api-v2/app/access/user.py
@@ -4,9 +4,9 @@ from fastapi import APIRouter, Request, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 
-from schemas import User
-from crud import get_user
-from api.deps import get_db   
+from app.access.schemas import User
+from app.access.crud import get_user
+from app.api.deps import get_db   
 
 
 router = APIRouter()
